Backend/requirement.txt
fastapi==0.115.12
pyyaml==6.0.2
dotenv==0.9.9
uvicorn==0.34.2
motor==3.7.1
colorama==0.4.6
bcrypt==4.3.0
python_jose==3.4.0